// src/models/user-infor.ts
export interface UserInfor {
  name: string;
  gender: string;
  age: string;
  weight: string;
  height: string;
  goal: string;
  experienceLevel: string;
  equipment: string[];
  weeklyFrequency: number;
  trainingDuration: number;
  priorityMuscleGroups: string[];
  restrictionsOrInjuries: string[];
  exercisePreferences: string[];
  progressGoal: string;
  specialConditions: string[];
  preferredIntensity: string;
  trainingLocation: string;
}